作品管理ページにおける「編集」ボタンが押下できないバグを修正
